[Patchset 3] trace_processor: add base trace loading code
This CL sets out the foundation for the trace loading code.
Specifically, it introduces BlobReader (an abstraction for reading data
from an arbitary data source into a buffer) and TraceLoader.

TraceLoader will use BlobReader and a class known as ProtoParser (to be
introduced in a follow-up CL) to read and parse the trace packets and
place them in the ColumnarTrace introduced in the previous CL

Bug: 80416541

+++ src/trace_processor/blob_reader.h
@@ -17,7 +17,7 @@
 #ifndef SRC_TRACE_PROCESSOR_BLOB_READER_H_
 #define SRC_TRACE_PROCESSOR_BLOB_READER_H_
 
-#include <cstdint>
+#include <stdint.h>
 
 namespace perfetto {
 namespace trace_processor {
@@ -28,7 +28,7 @@ class BlobReader {
   virtual ~BlobReader();
 
   // Reads |len| bytes at |offset| into |dst|.
-  virtual uint32_t Read(uint32_t offset, uint32_t len, uint8_t* dst) = 0;
+  virtual uint32_t Read(uint64_t offset, uint32_t len, uint8_t* dst) = 0;
 };
 
 }  // namespace trace_processor
